{cmake} Update LASFWFIO plugin
I cannot build this plugin, and the CI doesn't either, so these will require that someone try building it before merging.
