+++ plugins/core/IO/qLASFWFIO/include/LASFWFFilter.h
