+++ plugins/core/IO/qLASFWFIO/src/CMakeLists.txt
@@ -1,0 +1,6 @@
+
+target_sources( ${PROJECT_NAME}
+    PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/LASFWFFilter.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/qLASFWFIO.cpp
+)
